refactor: rename model folder to store
