+++ react-components/src/store/slices/detailSlice.ts
