+++ react-components/src/store/slices/photoCardSlice.ts
