+++ react-components/src/store/slices/photosSlice.ts
